Refactor

+++ app/element/led.cpp
@@ -101,7 +101,7 @@ void Led::refresh()
     if (isValid()) {
         std::bitset<4> indexBit;
         for (int i = 0; i < inputSize(); ++i) {
-            indexBit[i] = input(inputSize() - i - 1)->value();
+            indexBit[i] = static_cast<int>(input(inputSize() - i - 1)->value());
         }
         index = static_cast<int>(indexBit.to_ulong());
     }
@@ -192,7 +192,7 @@ void Led::setSkin(const bool defaultSkin, const QString &fileName)
     if (isValid()) {
         std::bitset<4> indexBit;
         for (int i = 0; i < inputSize(); ++i) {
-            indexBit[i] = input(inputSize() - i - 1)->value();
+            indexBit[i] = static_cast<int>(input(inputSize() - i - 1)->value());
         }
         index = static_cast<int>(indexBit.to_ulong());
     }
